Backend/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dateutils==0.6.12
Django==3.2.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==2.10
nsetools==1.0.11
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
wincertstore==0.2
